Add normalize indicator
For disparity map we don't need to obtain intensity between [0, 1].

Disparity map is a matrix of size that equals image size.
The elements of this matrix are shifts of pixels in left image to
the place where they are in the right image.

Left image is an image taken from left camera.
Right image is an image taken from right camera.