Implement function for deduplicating array
Function takes `array` of floats and `tolerance`.
Element is removed from array if it differs
from previous element in array not more than by `tolerance`.
Function is recursive as if we, for example,
delete the second element of array, we need to compare the first element with the third one,
but not the third with the fourth.