Implement function for epsilon search
An array of all possible epsilons
is created in `create_array_of_epsilons()`.
Array is sorted in ascendign order.
All duplicated from array are removed.

Firstly, epsilon is a median of the array.
If with this epsilon `crossing_out_graph` is not empty,
then we can try smaller epsilon (median of the array,
that is left from the current epsilon, or current median).
If the graph is empty,
then we find epsilon in the right part of the array.