Fix epsilon search
`epsilon_search()` was implemented as recursive binary search:
https://www.geeksforgeeks.org/binary-search/

Array is sorted in ascending order.

If with current epsilon (that is a median of current sub-array)
the problem is not epsilon-consistent,
then call the function with the left sub-array.
Else -- with the right sub-array.

When there is only one element in sub-array, the problem can be:
- epsilon-consistent, then return this element as epsilon
- not epsilon-consistent,
  then return the next element after the current element as epsilon